[chore] Move obs_queue to the internal/queue package
