+++ exporter/exporterhelper/internal/queue/obs_queue_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package queuebatch
+package queue
 
 import (
 	"context"
@@ -16,7 +16,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/metadatatest"
-	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/queue"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/requesttest"
 	"go.opentelemetry.io/collector/exporter/exportertest"
@@ -26,7 +25,7 @@ import (
 var exporterID = component.NewID(exportertest.NopType)
 
 type fakeQueue[T any] struct {
-	queue.Queue[T]
+	Queue[T]
 	offerErr error
 	size     int64
 	capacity int64
@@ -44,7 +43,7 @@ func (fq *fakeQueue[T]) Offer(context.Context, T) error {
 	return fq.offerErr
 }
 
-func newFakeQueue[T request.Request](offerErr error, size, capacity int64) queue.Queue[T] {
+func newFakeQueue[T request.Request](offerErr error, size, capacity int64) Queue[T] {
 	return &fakeQueue[T]{offerErr: offerErr, size: size, capacity: capacity}
 }
