+++ exporter/exporterhelper/internal/queue/queue.go
@@ -81,10 +81,24 @@ func (set *Settings[T]) activeSizer() request.Sizer[T] {
 	}
 }
 
-func NewQueue[T any](set Settings[T], next ConsumeFunc[T]) (Queue[T], error) {
+func NewQueue[T request.Request](set Settings[T], next ConsumeFunc[T]) (Queue[T], error) {
+	q, err := newBaseQueue(set)
+	if err != nil {
+		return nil, err
+	}
+
+	oq, err := newObsQueue(set, newAsyncQueue(q, set.NumConsumers, next))
+	if err != nil {
+		return nil, err
+	}
+
+	return oq, nil
+}
+
+func newBaseQueue[T any](set Settings[T]) (readableQueue[T], error) {
 	// Configure memory queue or persistent based on the config.
 	if !set.StorageID.HasValue() {
-		return newAsyncQueue(newMemoryQueue[T](set), set.NumConsumers, next), nil
+		return newMemoryQueue[T](set), nil
 	}
 	if set.ItemsSizer == nil {
 		return nil, errors.New("PersistentQueue requires ItemsSizer to be set")
@@ -92,7 +106,8 @@ func NewQueue[T any](set Settings[T], next ConsumeFunc[T]) (Queue[T], error) {
 	if set.BytesSizer == nil {
 		return nil, errors.New("PersistentQueue requires BytesSizer to be set")
 	}
-	return newAsyncQueue(newPersistentQueue[T](set), set.NumConsumers, next), nil
+
+	return newPersistentQueue[T](set), nil
 }
 
 // TODO: Investigate why linter "unused" fails if add a private "read" func on the Queue.
